pyproject.toml
[project]
name = "dino_loader"
version = "0.3.0"
description = "HPC-grade DINOv3 data pipeline for B200 / GB200 NVL72 clusters"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # ── Core ──────────────────────────────────────────────────────────────────
    # ~= pins the minor version: allows bugfix releases, blocks major changes.
    "torch~=2.10",
    "torchvision~=0.25",
    "numpy~=2.0",
    # ── WebDataset ────────────────────────────────────────────────────────────
    "webdataset~=1.0",
    # ── Async I/O ─────────────────────────────────────────────────────────────
    "aiofiles~=25.1",
    # ── Monitoring / CLI ──────────────────────────────────────────────────────
    "rich~=14.3",
    "psutil~=7.2",
    # ── Pillow: CPU backend augmentation pipeline ─────────────────────────────
    "pillow~=11.0",
    "dino-datasets",
    "dino-env",
    "nvidia-dali-cuda120>=2.0.0",
    "torchdata>=0.11.0",
]

[dependency-groups]
dev = [
    "pytest~=8.0",
    "pytest-timeout~=2.3",
    "pytest-mock~=3.14",
    "pytest-cov",
    "ruff",
    "ty",
]

# ── Optional GPU-only / cluster dependencies ──────────────────────────────────
# These cannot be in [dependencies] because they are unavailable on CPU-only
# machines (CI, laptops, macOS) and some cannot be resolved from PyPI directly.
#
# Install with:
#   pip install "dino_loader[gpu]"        # GPU training: DALI + TE
#   pip install "dino_loader[monitoring]" # Prometheus metrics endpoint
#   pip install "dino_loader[all]"        # Everything

[project.optional-dependencies]

# DALI: install separately with the correct CUDA index URL.
# See: https://docs.nvidia.com/deeplearning/dali/user-guide/docs/installation.html
# Do NOT add nvidia-dali to [dependencies] — it breaks CPU-only installs.
#
# Example:
#   pip install nvidia-dali-cuda120==1.40.*

gpu = [
    # Transformer Engine for FP8 quantisation (use_fp8_output=True).
    # [B4-FIX] Moved from [dependencies] to [optional-dependencies].
    # use_fp8_output defaults to False; LoaderConfig.__post_init__ raises a
    # clear error if FP8 is requested but TE is not installed.
    "transformer-engine~=2.12",
]

monitoring = [
    # Prometheus metrics endpoint (LoaderConfig.prometheus_port != None).
    # [ARCH3] Enables the opt-in Prometheus HTTP server.
    "prometheus-client~=0.21",
]

[tool.dino_loader.datasets]
# Override via env var DINO_DATASETS_ROOT or CLI --root flag.
# root = "/lustre/datasets"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: requires CUDA device",
    "dali: requires nvidia-dali",
    "slow: long-running integration tests",
]
timeout = 120
# Default: exclude slow tests.  Run `pytest` with no -m flag to get the fast
# suite.  Add `-m slow` or remove the addopts line to include slow tests.
addopts = "-m 'not slow'"

[tool.uv.sources]
dino-datasets = { git = "https://github.com/piflepaf1-eng/dino_datasets" }
dino-env = { git = "https://github.com/sebastienwood/dino_env" }

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN003",
    "D100",
    "D101",
    "D103",
    "D203",
    "D213",
    "INP001",
    "D105",
    "D400",
    "FBT001",
    "FBT002",
    "D415",
]

[tool.ruff.lint.per-file-ignores]
"*test*.py" = [
    "PLR2004",
    "S101",
    "ANN204",
    "ANN201",
    "ANN202",
    "ERA001",
    "D102",
    "ANN001",
    "PLC0415",
    "SLF001",
    "PT011",
    "D205",
    "ANN002",
    "D104",
    "PLR0913",
    "BLE001",
    "E402",
    "TRY003",
    "EM101",
    "D401",
]
"*scripts/*.py" = ["ALL"]